requirements-dev.txt
-e git+git@github.com:xmunoz/ipet-reader.git#egg=ipet-reader
backports-abc==0.4
click==6.6
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
flake8==3.2.1
mccabe==0.5.2
numpy==1.11.0
pandas==0.18.1
py==1.4.31
pycodestyle==2.2.0
pyflakes==1.3.0
pyparsing==2.1.5
pytest==3.0.2
python-dateutil==2.5.3
python-gitlab==0.14
pytz==2016.4
pyyaml==3.12
requests==2.10.0
requests-toolbelt==0.7.0
scipy==0.17.1
six==1.10.0
tornado==4.3
urllib3==1.15.1
